+++ kilde-assignment-api/src/main/java/com/senior/kilde/assignment/api/dto/InvestorCreateRequest.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class InvestorCreateRequest {
@@ -12,6 +14,6 @@ public class InvestorCreateRequest {
     private String name;
 
     @JsonProperty("initialBalanceAmount")
-    private Double initialBalanceAmount;
+    private BigDecimal initialBalanceAmount;
 
 }
