+++ kilde-assignment-api/src/main/java/com/senior/kilde/assignment/api/dto/TrancheCreateRequest.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class TrancheCreateRequest {
@@ -18,12 +20,12 @@ public class TrancheCreateRequest {
     private Integer duration;
 
     @JsonProperty("amountAvailableForInvestment")
-    private Double amountAvailableForInvestment;
+    private BigDecimal amountAvailableForInvestment;
 
     @JsonProperty("minimumInvestmentAmount")
-    private Double minimumInvestmentAmount;
+    private BigDecimal minimumInvestmentAmount;
 
     @JsonProperty("maximumInvestmentAmountPerInvestor")
-    private Double maximumInvestmentAmountPerInvestor;
+    private BigDecimal maximumInvestmentAmountPerInvestor;
 
 }
